Implement subcleanup thread
If the cleanup thread is not a daemon thread, program exit will
hang, waiting on the exit event, which never gets set because the
main thread is already gone. But making it a daemon thread means
execution may mid-cleanup, which may be problematic. This change
makes the cleanup thread a daemon, but when cleanup begins it
launches a second, non-daemon thread to perform the actual work,
and then waits for it to complete. This way if the program
terminates, the subcleanup should run to completion, but we avoid
long hangs.