Open temporary connection if necessary in _finddbmodule
Normally _finddbmodule runs immediately following pool initialization
and examines a connection from the pool to determine the dbmodule.
However, if poolmin/poolinit is 0, there won't be any connections in
the pool. So instead in this case we temporarily open a connection to
examine. This has the added benefit that pool creation will fail at
pool creation time, even with a pool size of 0, the same as it would
with a populated pool size.