Rework connection pool close to no longer wait on in-use connections
Previously, closing connection pool would wait on checked-out
connections. This frequently resulted in deadlocks in cases
like unhandled exceptions. This change updates the close methods
of connection pools and connections such that closing the pool
no longer waits on any outstanding connections. Instead, the
pool closes immediately, and the outstanding connection checks
the pool status on close. If the pool is already closed,
the close method of the connection now simply calls the close
method of the underlying database connection.